\name{DebianEvolData-package}
\alias{DebianEvolData}
\docType{package}
\encoding{UTF-8}
\title{Debian historical data extraction}
\description{
  Utility package for downloading and processing Debian packages
  metadata history.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\keyword{debian}
